Move test extensions to matching namespace
